requirements.txt
PyMuPDF
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
langchain 
langchain-community 
langchain-openai
python-dotenv
pydantic
duckduckgo-search